Correct TEAMIDs for a couple 1890s teams
1)  Changed 1891's AA "PHQ" franchise's teamid from "PH4" to "PHQ".
2)  Changed 1892-1899 NL "WAS" franchise's teamid from "WAS" to "WSN".
3)  Changed entries in Batting.CSV and Appearances.CSV in some instances
to accommodate this.

These changes were to address the fact that teamid "PH4" was being used by
the "PHA" franchise from 1882-1890 and franchise "PHQ" in 1891, and that
teamid "WAS" was associated with the "WAS" franchise from 1892-1899 and
the "WSN" franchise from 2005-2015.

I based these changes on feedback from the "baseball-databank" Yahoo
Group, where I asked about the apparent teamid duplications.